Fix compiler warnings, set -Werror by default
This enables -Werror by default and fixes all compiler warnings
on GCC 9.2.1.

+++ examples/src/app-hdf5-create.c
@@ -162,18 +162,18 @@ int main(int argc, char** argv)
         /* Create a new file using default properties. */
         file_id = H5Fcreate(targetfile, H5F_ACC_TRUNC, H5P_DEFAULT,
                             H5P_DEFAULT);
-        printf("H5Fcreate: %d\n", file_id);
+        printf("H5Fcreate: %ld\n", (long) file_id);
 
         /* Create the data space for the dataset. */
         dims[0] = 4;
         dims[1] = 6;
         dataspace_id = H5Screate_simple(2, dims, NULL);
-        printf("H5Screate_simple: %d\n", dataspace_id);
+        printf("H5Screate_simple: %ld\n", (long) dataspace_id);
 
         /* Create the dataset. */
         dataset_id = H5Dcreate2(file_id, "/dset", H5T_STD_I32BE, dataspace_id,
                                 H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
-        printf("H5Dcreate2: %d\n", dataset_id);
+        printf("H5Dcreate2: %ld\n", (long) dataset_id);
 
         if (write_data) {
             int i, j;
